Teoria dos números, busca binária e LCA

+++ codigo/matemagica/modinverse.cpp
@@ -4,7 +4,7 @@
 // Se b eh primo, basta fazer
 // a^(b-2)
 
-long long inv(long long a, long long b){
+ll inv(ll a, ll b){
 	if(1 < a) return b - inv(b % a, a) * b/a;
 	else return 1;
 }
